pyproject.toml
[project]
name = "money-stuff-scheduler"
version = "0.1.0"
description = "Daily Money Stuff newsletter translator and email sender"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.13",
    "langchain-openai>=0.3.0",
    "langgraph>=1.1.3",
    "feedparser>=6.0.0",
    "markdown>=3.7",
    "python-dotenv>=1.1.0",
    "apscheduler>=3.11.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
